NotesApp/backend/requirements.txt
fastapi
uvicorn
pydantic

sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
psycopg2-binary
email-validator